u-boot-updatehub-script: Rework script to handle bootcount
The script has been reworked to take bootcount into consideration
and handle the fallback. We did it without using the boot limit
feature so we avoid storing the 'altbootcmd' variable, thus
reducing the need of using the environment.

The rework, allowed to simplify it and make the code easier to
follow.

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>